requirements.txt
absl-py==0.15.0
astunparse==1.6.3
cachetools==5.2.0
charset-normalizer==2.1.1
clang==5.0
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
einops==0.5.0
fastprogress==1.0.3
filelock==3.8.0
flatbuffers==1.12
fonttools==4.38.0
gast==0.4.0
google-api-core==2.10.2
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery==1.28.2
google-cloud-bigquery-storage==1.1.2
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==1.3.3
googleapis-common-protos==1.56.4
grpcio==1.50.0
grpcio-status==1.48.2
h5py==3.1.0
huggingface-hub==0.10.1
idna==3.4
imageio==2.23.0
importlib-metadata==5.0.0
joblib==1.2.0
keras==2.6.0
keras-preprocessing==1.1.2
kiwisolver==1.4.4
lxml==4.9.1
markdown==3.4.1
matplotlib==3.6.1
mtcnn==0.1.1
networkx==2.8.8
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
pandas==1.3.5
pandas-datareader==0.9.0
pandas-gbq==0.13.3
pandas-profiling==1.4.1
pillow==9.2.0
protobuf==3.19.6
pyarrow==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydata-google-auth==1.4.0
pytz==2022.5
pywavelets==1.4.1
pyyaml==6.0
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
sacremoses==0.0.53
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.3
sentencepiece==0.1.97
six==1.15.0
sklearn==0.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-addons==0.16.1
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2022.10.10
tokenizers==0.12.1
torchsummary==1.5.1
tqdm==4.64.1
transformers==4.22.1
typeguard==2.13.3
typing-extensions==3.7.4.3
urllib3==1.26.12
werkzeug==2.2.2
wrapt==1.12.1
zipp==3.9.0